+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
-## 1.9.0
+
+## 1.9.2
+
+- Removed old setting for "one click creation" because is supported now on v11
+- Add new setting 'Default image icon for on creation of a note' for set a default image when creating a note on the canvas
+- Bug fix: issue with the width ratio use the pixel value instead multiply the width
+- Bug fix: https://github.com/p4535992/foundryvtt-pin-cushion/issues/76
+- Try to fix: https://github.com/p4535992/foundryvtt-pin-cushion/issues/52 and https://github.com/p4535992/foundryvtt-pin-cushion/issues/66
+
+
+## 1.9.0-1
 
 - Update to v11